setup.py: Set QA dependencies as extra requirements
Many users only want the fix scripts. They do not require gftools qa.
